+++ packages/browser/rollup.config.mjs
@@ -64,7 +64,7 @@ const plugins = (es5, noExternal) => [
         inject: false,
     }),
     babel({
-        extensions: ['.js', '.jsx', '.ts', '.tsx'],
+        extensions: ['.mjs', '.js', '.jsx', '.ts', '.tsx'],
         babelHelpers: 'bundled',
         plugins: [
             '@babel/plugin-transform-nullish-coalescing-operator',
