+++ packages/browser/src/autocapture-utils.ts
@@ -1,12 +1,11 @@
 import { AutocaptureConfig, Properties } from './types'
 import { each, entries } from './utils'
 
-import { isNullish, isString, isUndefined } from '@posthog/core'
+import { isNullish, isString, isUndefined, isArray } from '@posthog/core'
 import { logger } from './utils/logger'
 import { window } from './utils/globals'
 import { isDocumentFragment, isElementNode, isTag, isTextNode } from './utils/element-utils'
 import { includes, trim } from '@posthog/core'
-import { isArray } from './utils/type-utils'
 
 export function splitClassString(s: string): string[] {
     return s ? trim(s).split(/\s+/) : []
