+++ packages/browser/src/extensions/exception-autocapture/type-checking.ts
@@ -1,8 +1,7 @@
 // Portions of this file are derived from getsentry/sentry-javascript by Software, Inc. dba Sentry
 // Licensed under the MIT License
 
-import { isFunction, isNull, isUndefined } from '@posthog/core'
-import { isObject } from '../../utils/type-utils'
+import { isFunction, isNull, isUndefined, isObject } from '@posthog/core'
 
 export function isEvent(candidate: unknown): candidate is Event {
     return !isUndefined(Event) && isInstanceOf(candidate, Event)
