+++ packages/browser/src/extensions/replay/triggerMatching.ts
@@ -4,9 +4,8 @@ import {
 } from '../../constants'
 import { PostHog } from '../../posthog-core'
 import { FlagVariant, RemoteConfig, SessionRecordingUrlTrigger } from '../../types'
-import { isNullish, isBoolean, isString } from '@posthog/core'
+import { isNullish, isBoolean, isString, isObject } from '@posthog/core'
 import { window } from '../../utils/globals'
-import { isObject } from '../../utils/type-utils'
 
 export const DISABLED = 'disabled'
 export const SAMPLED = 'sampled'
