+++ packages/browser/src/scroll-manager.ts
@@ -1,7 +1,7 @@
 import { window } from './utils/globals'
 import { PostHog } from './posthog-core'
 import { addEventListener } from './utils'
-import { isArray } from './utils/type-utils'
+import { isArray } from '@posthog/core'
 
 export interface ScrollContext {
     // scroll is how far down the page the user has scrolled,
