+++ packages/browser/src/sessionid.ts
@@ -7,10 +7,9 @@ import { window } from './utils/globals'
 
 import { createLogger } from './utils/logger'
 
-import { isNumber, isUndefined, clampToRange } from '@posthog/core'
+import { isArray, isNumber, isUndefined, clampToRange } from '@posthog/core'
 import { PostHog } from './posthog-core'
 import { addEventListener } from './utils'
-import { isArray } from './utils/type-utils'
 
 const logger = createLogger('[SessionId]')
