+++ packages/browser/src/utils/request-utils.ts
@@ -1,9 +1,8 @@
 import { each } from './'
 
-import { isFile, isUndefined } from '@posthog/core'
+import { isArray, isFile, isUndefined } from '@posthog/core'
 import { logger } from './logger'
 import { document } from './globals'
-import { isArray } from './type-utils'
 
 const localDomains = ['localhost', '127.0.0.1']
